third_party/blink/web_tests/editing/execCommand/justify_center_do_not_modify_root_editable.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '|centered text',
    '</div>',
  ],
  'JustifyCenter',
  [
    '<div contenteditable>',
      '<div style="text-align: center;">',
        '|centered text',
      '</div>',
    '</div>'
  ]),
  'JustifyCenter command do not modify its block rootEditableElement');
</script>
